# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.LoginResponse:
    properties:
      access_token:
        description: '#nosec G101,G117 -- This is a DTO field, not a hardcoded credential'
        type: string
      refresh_token:
        description: '#nosec G101,G117 -- This is a DTO field, not a hardcoded credential'
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.TwoFADisableRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  dto.TwoFAEnableResponse:
    properties:
      message:
        type: string
      recovery_codes:
        items:
          type: string
        type: array
    type: object
  dto.TwoFALoginRequest:
    properties:
      code:
        type: string
      recovery_code:
        type: string
      temp_token:
        type: string
    required:
    - temp_token
    type: object
  dto.TwoFAMethodsResponse:
    properties:
      available_methods:
        items:
          type: string
        type: array
      email_2fa_enabled:
        type: boolean
      passkey_enabled:
        type: boolean
      totp_enabled:
        type: boolean
    type: object
  dto.TwoFARecoveryCodesResponse:
    properties:
      message:
        type: string
      recovery_codes:
        items:
          type: string
        type: array
    type: object
...
  dto.TwoFAVerifyRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  twofa.TwoFASetupResponse:
    properties:
      qr_code_data:
        items:
          type: integer
        type: array
      qr_code_url:
        type: string
      secret:
        description: '#nosec G101,G117 -- This is a response field for TOTP secret,
          not a hardcoded credential'
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample authentication and authorization API built with Go
    and Gin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Authentication and Authorization API
  version: 1.1.0
paths:
  /2fa/disable:
    post:
      consumes:
      - application/json
      description: Disable 2FA for the user (requires TOTP code or email 2FA code)
      parameters:
      - description: Disable 2FA Data
        in: body
        name: disable
        required: true
        schema:
          $ref: '#/definitions/dto.TwoFADisableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disable 2FA
      tags:
      - 2FA
  /2fa/email/enable:
    post:
      description: Enable email-based 2FA for the user (sends codes to registered
        email)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TwoFAEnableResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Enable email-based 2FA
      tags:
      - 2FA
  /2fa/email/resend:
    post:
      consumes:
      - application/json
      description: Resend a new 2FA verification code to the user's email during login
      parameters:
      - description: Temporary login token
        in: body
        name: resend
        required: true
        schema:
          properties:
            temp_token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Resend email 2FA code
      tags:
      - 2FA
  /2fa/enable:
    post:
      description: Enable 2FA for the user after successful verification
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TwoFAEnableResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Enable 2FA
      tags:
      - 2FA
  /2fa/generate:
    post:
      description: Generate a 2FA secret and QR code for user setup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/twofa.TwoFASetupResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate 2FA setup
      tags:
      - 2FA
  /2fa/login-verify:
    post:
      consumes:
      - application/json
      description: Verify the 2FA code during the login process (supports TOTP, email
        code, or recovery code)
      parameters:
      - description: 2FA Login Data
        in: body
        name: verify
        required: true
        schema:
          $ref: '#/definitions/dto.TwoFALoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Verify 2FA during login
      tags:
      - 2FA
  /2fa/methods:
    get:
      description: Get the 2FA methods available for the current application
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TwoFAMethodsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get available 2FA methods
      tags:
      - 2FA
...
  /2fa/recovery-codes:
    post:
      consumes:
      - application/json
      description: Generate and display new recovery codes (requires password and/or
        TOTP code)
      parameters:
      - description: TOTP Code
        in: body
        name: verify
        required: true
        schema:
          $ref: '#/definitions/dto.TwoFAVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TwoFARecoveryCodesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate new recovery codes
      tags:
      - 2FA
  /2fa/verify-setup:
    post:
      consumes:
      - application/json
      description: Verify the initial 2FA setup with a TOTP code
      parameters:
      - description: TOTP Code
        in: body
        name: setup
        required: true
        schema:
          $ref: '#/definitions/dto.TwoFAVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Verify 2FA setup
      tags:
      - 2FA
...
securityDefinitions:
  AdminApiKey:
    description: Admin API Key for protected admin routes
    in: header
    name: X-Admin-API-Key
    type: apiKey
  ApiKeyAuth:
    description: Type "Bearer" + your JWT token
    in: header
    name: Authorization
    type: apiKey
  AppApiKey:
    description: Per-application API Key for app-scoped routes
    in: header
    name: X-App-API-Key
    type: apiKey
  AppID:
    description: The UUID of the application context
    in: header
    name: X-App-ID
    type: apiKey
swagger: "2.0"